# swagger/community-api.yaml
swagger: '2.0'
info:
  version: 0.9.0
  title: Västtrafik Developer Community API
  description: Community API for the developer portal
consumes:
  - application/json
produces:
  - application/json
schemes:
  - https
paths:
...
  /posts:
    get:
      tags:
        - post
      description: 'Retrieves posts, without edits'
      parameters:
        - name: label
          in: query
          description: Only retrieve posts with this label
          required: false
          type: string
          enum:
            - popular
            - recent
            - unanswered
            - votes
        - name: query
          in: query
          description: 'Query string that will be matched against the following attributes: createDate, createdBy, text, categoryId and forumId'
          required: false
          type: string
        - name: offset
          $ref: '#/parameters/offset'
        - name: limit
          $ref: '#/parameters/limit'
      responses:
        '200':
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/Post'
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
    post:
      tags:
        - post
      description: Creates a new post
      parameters:
        - $ref: '#/parameters/X-JWT-Assertion'
        - name: body
          in: body
          description: The post to create
          required: true
          schema:
            $ref: '#/definitions/Post'
      responses:
        '201':
          description: Successful response with the newly created post
          schema:
            $ref: '#/definitions/Post'
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/Error'
        '401':
          description: Unauthorized
        '404':
          description: Not found
        '500':
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
  '/posts/{id}':
    get:
      tags:
        - post
      description: 'Retrieves a post, including all edits'
      parameters:
        - $ref: '#/parameters/id'
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/Post'
        '404':
          description: Not found
        '500':
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
    put:
      tags:
        - post
      description: Updates a post
      parameters:
        - $ref: '#/parameters/X-JWT-Assertion'
        - $ref: '#/parameters/id'
        - name: action
          in: query
          description: An action parameter specifying the patch to perform
          type: string
          enum:
            - answered
            - edited
        - name: body
          in: body
          description: A post object
          required: true
          schema:
            $ref: '#/definitions/Post'
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/Post'
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/Error'
        '401':
          description: Unauthorized
        '404':
          description: Not found
        '500':
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
    delete:
      tags:
        - post
      description: 'Deletes a post, including all edits'
      parameters:
        - $ref: '#/parameters/X-JWT-Assertion'
        - $ref: '#/parameters/id'
      responses:
        '200':
          description: Successful response
        '401':
          description: Unauthorized
        '404':
          description: Not found
        '500':
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
  '/posts/{id}/edits':
    get:
      tags:
        - post
      description: 'Retrieves all versions of the post, sorted in descending order'
      parameters:
        - $ref: '#/parameters/id'
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/PostEdit'
        '404':
          description: Not found
        '500':
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
  '/posts/{id}/votes':
    post:
      tags:
        - vote
      description: Creates a vote
      parameters:
        - $ref: '#/parameters/X-JWT-Assertion'
        - $ref: '#/parameters/id'
        - name: body
          in: body
          description: The vote to create
          schema:
            $ref: '#/definitions/Vote'
      responses:
        '201':
          description: Successful response
          schema:
            $ref: '#/definitions/Vote'
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/Error'
        '401':
          description: Unauthorized
        '404':
          description: Not found
        '500':
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
...
parameters:
  X-JWT-Assertion:
    name: X-JWT-Assertion
    in: header
    description: JWT token
    required: true
    type: string
...
  id:
    name: id
    in: path
    description: Resource id
    required: true
    type: integer
    format: int64
  limit:
    name: limit
    in: query
    description: Maximum size array to return
    default: 10
    required: false
    type: integer
    format: int32
  offset:
    name: offset
    in: query
    description: Starting point of the list
    default: 1
    required: false
    type: integer
    format: int32
definitions:
...
  Attachment:
    description: Attachment to a message
    required:
      - messageId
      - fileName
      - mimeType
      - size
    properties:
      id:
        description: Attachment id
        type: integer
        format: int64
      messageId:
        description: The message this attachment belongs to
        type: integer
        format: int64
      fileName:
        description: The file name
        type: string
      mimeType:
        description: Mime type
        type: string
      size:
        description: Attachment size in bytes
        type: integer
        format: int64
...
  Error:
    title: Error
    description: Error message
    required:
      - code
      - message
    properties:
      code:
        description: Error code
        type: string
      message:
        description: Error message
        type: string
      description:
        description: Detailed error description
        type: string
      moreInfo:
        description: 'More information about the error, such as a URL'
        type: string
      items:
        description: 'If request resulted in multiple errors, list them in this array'
        type: array
        items:
          $ref: '#/definitions/ErrorListItem'
  ErrorListItem:
    title: ErrorListItem
    description: Description of individual errors that may have occored during a request
    required:
      - code
      - message
    properties:
      code:
        description: Error code
        type: string
      message:
        description: Error message
        type: string
...
  Member:
    title: Member
    description: A community member (user)
    required:
      - id
      - userName
      - email
    properties:
      id:
        description: Member id
        type: integer
        format: int32
      userName:
        description: Username
        type: string
      email:
        description: User email
        type: string
      status:
        description: The status of the user
        type: string
        enum:
          - inactive
          - active
          - probation
          - banned
          - deleted
      showEmail:
        description: Whether the email of the user should be displayed
        type: boolean
      signature:
        description: Signature to use
        type: string
      useGravatar:
        description: Whether to use Gravatar
        type: boolean
      gravatarEmail:
        description: The Gravatar email of the user
        type: string
      acceptAllMessages:
        description: Whether private messages is enabled
        type: boolean
      notifyEmail:
        description: Whether the user should be notified by email
        type: boolean
      notifyMessage:
        description: Whether the user should be notified by internal message
        type: boolean
      notifyText:
        description: Whether the text should be included in notifications
        type: boolean
      showRankings:
        description: Whether ranking information should be displayed
        type: boolean
      rankings:
        description: The user rankings
        type: array
        items:
          $ref: '#/definitions/MemberRanking'
  MemberRanking:
    title: MemberRanking
    description: Ranking for a community member
    required:
      - title
      - type
      - imageURL
    properties:
      id:
        description: Ranking id
        type: integer
        format: int32
      memberId:
        description: Member id
        type: integer
        format: int32
      title:
        description: Ranking title
        type: string
      type:
        description: Ranking type
        type: string
        enum:
          - poster
          - reputation
      imageURL:
        description: A ranking image
        type: string
      minPoints:
        description: The number of points required for this ranking
        type: integer
        format: int32
      currentPoints:
        description: The number of points currently held
  Message:
    title: Message
    description: A private message between community member or between admins and community members
    required:
      - folderId
      - subject
      - body
      - sender
      - recipients
    properties:
      id:
        description: Message id
        type: integer
        format: int64
      originalId:
        description: 'The original message id, in case this message is a reply or forward.'
        type: integer
        format: int64
      folderId:
        description: The folder this message belongs to
        type: integer
        format: int64
      subject:
        description: 'The subject of the message. In case the message is a reply or a forward, the subject will reflect this.'
        type: string
      body:
        description: The text of the message
        type: string
      htmlEnabled:
        description: Whether html rendering is enabled
        type: boolean
        default: true
      createDate:
        description: Date and time when the message was created
        type: string
        format: date-time
      readDate:
        description: The date and time this message was opened for reading by the recipient
        type: string
        format: date-time
      sender:
        description: The user that created the message
        $ref: '#/definitions/Member'
      recipients:
        description: The recipients of the message
        type: array
        items:
          $ref: '#/definitions/Member'
      numberOfAttachments:
        description: Number of attachments in the message
        type: integer
        format: int32
      attachments:
        description: Attachments
        type: array
        items:
          $ref: '#/definitions/Attachment'
  Post:
    title: Post
    description: 'A post belongs to a topic and can be a question, answer, comment or report'
    required:
      - topicId
      - type
      - text
      - textFormat
    properties:
      id:
        description: Post id
        type: integer
        format: int64
      topicId:
        description: The topic this post was submitted to
        type: integer
        format: int64
      forumId:
        description: The forum this post was submitted to
        type: integer
        format: int64
      categoryId:
        description: The category this post was submitted to
        type: integer
        format: int64
      type:
        description: The post type
        type: string
        enum:
          - question
          - answer
          - comment
          - report
      text:
        description: The text of the post
        type: string
      textFormat:
        description: 'Markup language, or plain text'
        type: string
        enum:
          - hmtl
          - md
          - plain
      createDate:
        description: Date and time when the post was submitted
        type: string
        format: date-time
      createdBy:
        description: The user that submitted the post
        $ref: '#/definitions/Member'
      commentTo:
        description: If the post is a reply to some other post
        $ref: '#/definitions/Post'
      pointsAwarded:
        description: Points awarded by votes
        type: integer
        format: int32
      isAnswer:
        description: If this post has been selected as the answer to the topic question
        type: boolean
      editDate:
        description: The date and time the post was last edited
        type: string
        format: date-time
      editedBy:
        description: The user that performed the last edit
        $ref: '#/definitions/Member'
      numberOfTimesEdited:
        description: The number of times the post has been edited
        type: integer
        format: int32
      edits:
        description: 'Previous versions of the post, if it has been edited'
        type: array
        items:
          $ref: '#/definitions/PostEdit'
      reports:
        description: 'If the post has been reported, an array of all the reports'
        type: array
        items:
          $ref: '#/definitions/Report'
      isModerated:
        description: If the post has been moderated. A moderated post may only be edited by a moderator.
        type: boolean
        default: false
      isDeleted:
        description: If this post has been deleted (flagged as deleted)
        type: boolean
        default: false
      isReported:
        description: If the post has been reported
        type: boolean
        default: false
  PostEdit:
    title: PostEdit
    description: A previus version of a post that has been edited
    required:
      - postId
      - version
      - text
      - ml
    properties:
      id:
        description: Post edit id
        type: integer
        format: int64
      postId:
        description: The post this edit belongs to
        type: integer
        format: int64
      version:
        description: The post version
        type: integer
        format: int32
      text:
        description: The text of the post
        type: string
      ml:
        description: Markup language
        type: string
        enum:
          - hmtl
          - md
      createDate:
        description: Date and time when the post was submitted
        type: string
        format: date-time
      createdBy:
        description: The user that submitted the post
        $ref: '#/definitions/Member'
...
  Report:
    title: Report
    description: 'A report can be submitted by a community member in response to a post containing obscene language, disrespectful remarks etc, but can also be automatically generated by the system as a result of text analysis'
    required:
      - id
      - post
      - type
      - text
    properties:
      id:
        description: Report id
        type: integer
        format: int64
      post:
        description: The post that was reported
        $ref: '#/definitions/Post'
      type:
        description: The report type
        type: string
        enum:
          - user
          - system
      text:
        description: The report text
        type: string
      reportDate:
        description: The date and time the report was submitted
        type: string
        format: date-time
      reportedBy:
        description: The user that submitted the report
        $ref: '#/definitions/Member'
      measures:
        description: The measures having been taken in response to this report
        type: array
        items:
          $ref: '#/definitions/ReportMeasure'
  ReportMeasure:
    title: ReportMeasure
    description: Measure taken in response to a report
    required:
      - id
      - type
    properties:
      id:
        description: Measure id
        type: integer
        format: int64
      type:
        description: The measure type
        type: string
        enum:
          - asked
          - warned
          - banned
          - none
      message:
        description: The message sent to the user having been reported
        $ref: '#/definitions/Message'
      measureDate:
        description: The date and time the measure was taken
        type: string
        format: date-time
      rectifiedBy:
        description: The user that took measure in response to the report
        $ref: '#/definitions/Member'
      reporterNotification:
        description: The message sent to the user submitting the report
        $ref: '#/definitions/Message'
...
  Vote:
    title: Vote
    description: A vote is a way for community members to award points for questions and answers. The points are then used for ranking purposes.
    required:
      - postId
      - type
      - memberId
      - points
    properties:
      id:
        description: User id
        type: integer
        format: int32
      topicId:
        description: The topic id
        type: integer
        format: int64
      postId:
        description: The post id
        type: integer
        format: int64
      type:
        description: The voting type
        type: string
        enum:
          - question
          - answer
      memberId:
        description: The user casting the vote
        type: integer
        format: int64
      points:
        description: Number of points awarded
        type: integer
        format: int32
...